Add files via upload
Additional versions 1119, 1234, and 1331 of the TSS have been released to the TSS sourceforge site since the Platform Verification tool set was released.  Using TSS version 1119, the tool set will compile without a problem. Using TSS versions 1234 and 1331 will show errors.  This happens because there was a change in the directory structure of the TSS source code in the latter versions. Check to make sure that the local directory structure is setup correctly.