Add files via upload
Additional versions 1119, 1234, and 1331of the TSS have been released to the TSS sourceforge site since the Platform Verify tool set was released. Using Version 1119, the tool set will compile without a problem, but using versions 1234 and 1331 will show errors. This happens because there was a change in the directory structure of the TSS source code in the latter versions. Make sure the local directory structure is set up correctly.